Add preliminary flower object
Currently displays as a purple capsule, pending creation of proper assets

Every 5 seconds, it creates a new nectar instance (for now the "PickUp"
game object) somewhere nearby, for the player (bee) to try and capture.
This replaces the 16 initial instances in the scene, which are removed.

Additionally, when the player collides with a nectar, destroy it instead of
hiding it.