Apply Bee display to the player
For now, just static display of the bee, no rotation as the object changes
direction.

Additionally, freeze the rotation of the underlying sphere in the player
object, so that the bee doesn't rotate around like crazy. I'll likely add
different input handling later to adjust the direction that the bee is facing
to match the movement.