Bee: look in the direction of the last arrow press
Whenever the movement is updated, if currently moving (i.e. not from letting go
of an arrow) have the bee face the direction that was last pressed. This is not
exactly the same as the direction of travel, since the bee will maintain any
momentum, but is reflective of the direction the player wants to move the bee.