Bee: bring nectar to the hive
Rather than counting collected nectar immediately, require the bee to visit the
hive to drop it off first. The inventory of the bee (the number of nectar
objects currently being carried) is not currently shown to the user, but it
likely will be displayed in a subsequent commit.

Additionally, tag the PlayerController with the "Player" tag to allow
detecting the collisions that trigger dropping off the nectar inventory.