Add PinkFlowerObj prefab
Instead of just including a bunch of pink flower displays, create a prefab
object that includes both the display and the script to actually create nectar
objects.

Additionally, tweak Flower.cs to ensure that the nectars are always created at
y=0.6, regardless of the position of the flower itself, so that they can be
collected by the bee.

Will be followed by similar commits for the orange and white flowers.